Fixed tool agent inheritance
