+++ camel/agents/__init__.py
@@ -11,8 +11,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
-from .base import BaseAgent
-from .chat_agent import ChatAgent, ChatAgentResponse
+from .base import BaseAgent, ChatAgentResponse
+from .chat_agent import ChatAgent
 from .task_agent import TaskPlannerAgent, TaskSpecifyAgent
 from .critic_agent import CriticAgent
 from .tool_agents.base import BaseToolAgent
